UI Updates, SSR vs. CSR, and dnssec-prover updates (#18)
* removed email field

* updated dnssec-prover

* check for duplicate record

* moved footer into layout

* moved create form into feature component

* moved check form into feature component

* tagline feature component and ssr for homepage

* pull default domain from env
