+++ src/app/layout.tsx
@@ -16,7 +16,17 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <div className="text-purple-800 bg-12teal h-full flex flex-col justify-between">
+          {children}
+          <footer className="border-t border-t-purple-800 pt-2 mt-4 text-purple-800">
+            <ul className="flex flex-col gap-2 md:flex-row md:gap-4 md: p-4">
+              <li><a href="https://github.com/ATLBitLab/twelvecash">Developer Docs</a></li>
+              <li>Made with  🧡 at <a href="https://atlbitlab.com/">ATL BitLab</a></li>
+            </ul>
+          </footer>
+        </div>
+      </body>
     </html>
   )
 }
