+++ src/lib/dnssec-prover/dnssec_prover_wasm.js
@@ -181,26 +181,30 @@ export function get_unverified_proof(proof_builder) {
 }
 
 /**
-* Verifies an RFC 9102-formatted proof and returns the [`VerifiedRRStream`] in JSON form.
+* Verifies an RFC 9102-formatted proof and returns verified records matching the given name
+* (resolving any C/DNAMEs as required).
 * @param {Uint8Array} stream
+* @param {string} name_to_resolve
 * @returns {string}
 */
-export function verify_byte_stream(stream) {
-    let deferred2_0;
-    let deferred2_1;
+export function verify_byte_stream(stream, name_to_resolve) {
+    let deferred3_0;
+    let deferred3_1;
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         const ptr0 = passArray8ToWasm0(stream, wasm.__wbindgen_malloc);
         const len0 = WASM_VECTOR_LEN;
-        wasm.verify_byte_stream(retptr, ptr0, len0);
+        const ptr1 = passStringToWasm0(name_to_resolve, wasm.__wbindgen_malloc, wasm.__wbindgen_realloc);
+        const len1 = WASM_VECTOR_LEN;
+        wasm.verify_byte_stream(retptr, ptr0, len0, ptr1, len1);
         var r0 = getInt32Memory0()[retptr / 4 + 0];
         var r1 = getInt32Memory0()[retptr / 4 + 1];
-        deferred2_0 = r0;
-        deferred2_1 = r1;
+        deferred3_0 = r0;
+        deferred3_1 = r1;
         return getStringFromWasm0(r0, r1);
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
-        wasm.__wbindgen_free(deferred2_0, deferred2_1, 1);
+        wasm.__wbindgen_free(deferred3_0, deferred3_1, 1);
     }
 }
