+++ src/lib/dnssec-prover/dnssec_prover_wasm_bg.wasm.d.ts
@@ -6,7 +6,7 @@ export function init_proof_builder(a: number, b: number, c: number): number;
 export function process_query_response(a: number, b: number, c: number): void;
 export function get_next_query(a: number, b: number): void;
 export function get_unverified_proof(a: number, b: number): void;
-export function verify_byte_stream(a: number, b: number, c: number): void;
+export function verify_byte_stream(a: number, b: number, c: number, d: number, e: number): void;
 export function __wbindgen_malloc(a: number, b: number): number;
 export function __wbindgen_realloc(a: number, b: number, c: number, d: number): number;
 export function __wbindgen_add_to_stack_pointer(a: number): number;
